Fix buffer life-cycle events
Make the BufReadCmd a ++nested autocommand so that all relevant events
are triggered (important for external plugins listening to buffer events
via autocommands).

Also fire the BufWinEnter via doautocmd after BufRead. This is the next
event in the buffer life-cycle that is fired (:help BufRead and its
reference to BufWinEnter)